feat: CommunityBook 평점 조회 기능 구현

+++ autobiography/src/main/java/com/c203/autobiography/domain/communityBook/dto/CommunityBookRatingRequest.java
@@ -10,6 +10,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 // 평점 생성 요청 DTO
 @Getter
 @NoArgsConstructor
@@ -27,5 +29,5 @@ public class CommunityBookRatingRequest {
     @Min(value = 1, message = "평점은 1점 이상이어야 합니다.")
     @Max(value = 5, message = "평점은 5점 이하여야 합니다.")
     @Schema(description = "평점 (1-5점)", example = "4", required = true)
-    private Integer score;
+    private BigDecimal  score;
 }
